Feat. add useMutateDeleteDocument
